Finishing Admin Dashboard Tags

+++ app/Http/Controllers/AdminControllers/AdminCategoriesController.php
@@ -64,6 +64,10 @@ class AdminCategoriesController extends Controller
     public function destroy(Category $category)
     {
         $default_category_id = Category::where('name','Uncategorized')->first()->id;
+
+        if($category->name ==='Uncategorized')
+            abort(404);
+            
         $category->posts()->update(['category_id'=>$default_category_id]);
         
         $category->delete();
